+++ previous-versions/assess/pkg/models/operations/getaccountingmarketingmetrics.go
@@ -8,7 +8,9 @@ import (
 )
 
 type GetAccountingMarketingMetricsRequest struct {
-	CompanyID    string `pathParam:"style=simple,explode=false,name=companyId"`
+	// Unique identifier for a company.
+	CompanyID string `pathParam:"style=simple,explode=false,name=companyId"`
+	// Unique identifier for a connection.
 	ConnectionID string `pathParam:"style=simple,explode=false,name=connectionId"`
 	// Shows the dimensionDisplayName and itemDisplayName in measures to make the report data human-readable.
 	IncludeDisplayNames *bool `queryParam:"style=form,explode=true,name=includeDisplayNames"`
@@ -81,12 +83,15 @@ func (o *GetAccountingMarketingMetricsRequest) GetShowInputValues() *bool {
 }
 
 type GetAccountingMarketingMetricsResponse struct {
+	// HTTP response content type for this operation
 	ContentType string
 	// Your API request was not properly authorized.
 	ErrorMessage *shared.ErrorMessage
 	// OK
-	Report      *shared.Report
-	StatusCode  int
+	Report *shared.Report
+	// HTTP response status code for this operation
+	StatusCode int
+	// Raw HTTP response; suitable for custom response parsing
 	RawResponse *http.Response
 }