+++ previous-versions/assess/pkg/models/operations/getdataintegritystatus.go
@@ -8,6 +8,7 @@ import (
 )
 
 type GetDataIntegrityStatusRequest struct {
+	// Unique identifier for a company.
 	CompanyID string `pathParam:"style=simple,explode=false,name=companyId"`
 	// A key for a Codat data type.
 	DataType shared.DataIntegrityDataType `pathParam:"style=simple,explode=false,name=dataType"`
@@ -28,12 +29,15 @@ func (o *GetDataIntegrityStatusRequest) GetDataType() shared.DataIntegrityDataTy
 }
 
 type GetDataIntegrityStatusResponse struct {
+	// HTTP response content type for this operation
 	ContentType string
 	// Your API request was not properly authorized.
 	ErrorMessage *shared.ErrorMessage
 	// OK
-	Status      *shared.Status
-	StatusCode  int
+	Status *shared.Status
+	// HTTP response status code for this operation
+	StatusCode int
+	// Raw HTTP response; suitable for custom response parsing
 	RawResponse *http.Response
 }