+++ previous-versions/assess/pkg/models/shared/payment.go
@@ -3,11 +3,13 @@
 package shared
 
 import (
-	"github.com/codatio/client-sdk-go/previous-versions/assess/pkg/types"
+	"github.com/codatio/client-sdk-go/previous-versions/assess/pkg/utils"
+	"github.com/ericlagergren/decimal"
 )
 
 type Payment struct {
-	Amount                 *types.Decimal          `json:"amount,omitempty"`
+	// Payment amount.
+	Amount                 *decimal.Big            `decimal:"number" json:"amount,omitempty"`
 	BankingTransactionRefs []BankingTransactionRef `json:"bankingTransactionRefs,omitempty"`
 	// The currency data type in Codat is the [ISO 4217](https://en.wikipedia.org/wiki/ISO_4217) currency code, e.g. _GBP_.
 	//
@@ -42,7 +44,7 @@ type Payment struct {
 	// | **GBP**          | £20            | 1.277         | $25.54                     |
 	// | **EUR**          | €20            | 1.134         | $22.68                     |
 	// | **RUB**          | ₽20            | 0.015         | $0.30                      |
-	CurrencyRate *types.Decimal `json:"currencyRate,omitempty"`
+	CurrencyRate *decimal.Big `decimal:"number" json:"currencyRate,omitempty"`
 	// In Codat's data model, dates and times are represented using the <a class="external" href="https://en.wikipedia.org/wiki/ISO_8601" target="_blank">ISO 8601 standard</a>. Date and time fields are formatted as strings; for example:
 	//
 	// ```
@@ -63,12 +65,24 @@ type Payment struct {
 	// > Not all dates from Codat will contain information about time zones.
 	// > Where it is not available from the underlying platform, Codat will return these as times local to the business whose data has been synced.
 	Date *string `json:"date,omitempty"`
-	// ID of the invoice, which may be a GUID but it may be something else depending on the accounting platform
-	ID          *string `json:"id,omitempty"`
+	// ID of the invoice, which may be a GUID but it may be something else depending on the accounting platform.
+	ID *string `json:"id,omitempty"`
+	// The type of payment.
 	PaymentType *string `json:"paymentType,omitempty"`
 }
 
-func (o *Payment) GetAmount() *types.Decimal {
+func (p Payment) MarshalJSON() ([]byte, error) {
+	return utils.MarshalJSON(p, "", false)
+}
+
+func (p *Payment) UnmarshalJSON(data []byte) error {
+	if err := utils.UnmarshalJSON(data, &p, "", false, false); err != nil {
+		return err
+	}
+	return nil
+}
+
+func (o *Payment) GetAmount() *decimal.Big {
 	if o == nil {
 		return nil
 	}
@@ -89,7 +103,7 @@ func (o *Payment) GetCurrency() *string {
 	return o.Currency
 }
 
-func (o *Payment) GetCurrencyRate() *types.Decimal {
+func (o *Payment) GetCurrencyRate() *decimal.Big {
 	if o == nil {
 		return nil
 	}